backend/requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asyncmy==0.2.7
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
distlib==0.3.6
exceptiongroup==1.1.1
fastapi==0.95.0
filelock==3.10.7
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
Jinja2==3.1.2
limits==2.8.0
MarkupSafe==2.1.2
packaging==22.0
pipenv==2023.3.20
platformdirs==3.2.0
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.7
pypika-tortoise==0.1.6
pytest==7.2.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
slowapi==0.1.7
sniffio==1.3.0
sse-starlette==1.3.3
starlette==0.26.1
tomli==2.0.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
watchfiles==0.19.0
websockets==10.4
wrapt==1.15.0
youtube-transcript-api==0.5.0
